// src/LupiraCalWeb.Client/src/data/api/models/itemKind.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * LupiraCalApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type ItemKind = typeof ItemKind[keyof typeof ItemKind];


export const ItemKind = {
  Generic: 'Generic',
  Travel: 'Travel',
  Flight: 'Flight',
  Train: 'Train',
  Bus: 'Bus',
  Car: 'Car',
  Lodging: 'Lodging',
  Appointment: 'Appointment',
  Ticketed: 'Ticketed',
  Delivery: 'Delivery',
  Bill: 'Bill',
  Availability: 'Availability',
} as const;